refactor: extract incremental graph to separate file
